fix: move rounds_info to decision maker
